move shutdown functions in EntryPoint
